Add variables for storing settings and submittedContracts in LocalStorage
