+++ apps/contract-verification/src/app/Verifiers/AbstractVerifier.ts
@@ -1,5 +1,5 @@
 import { CompilerAbstract } from '@remix-project/remix-solidity'
-import { LookupResponse, SubmittedContract, VerificationResponse, VerificationStatus } from '../types/VerificationTypes'
+import type { LookupResponse, SubmittedContract, VerificationResponse, VerificationStatus } from '../types'
 
 export interface AbstractVerifier {
   checkVerificationStatus?(receiptId: string): Promise<VerificationStatus>
