+++ apps/contract-verification/src/app/Verifiers/EtherscanVerifier.ts
@@ -1,6 +1,6 @@
 import { CompilerAbstract } from '@remix-project/remix-solidity'
 import { AbstractVerifier } from './AbstractVerifier'
-import { LookupResponse, SubmittedContract, VerificationResponse, VerificationStatus } from '../types/VerificationTypes'
+import type { LookupResponse, SubmittedContract, VerificationResponse, VerificationStatus } from '../types'
 
 interface EtherscanVerificationRequest {
   chainId?: string
