+++ apps/contract-verification/src/app/types/SettingsTypes.ts
@@ -1,0 +1,19 @@
+import { VerifierIdentifier } from './VerificationTypes'
+
+export interface VerifierSettings {
+  apiUrl: string
+  explorerUrl?: string
+  apiKey?: string
+}
+
+export type SettingsForVerifier = Record<VerifierIdentifier, VerifierSettings>
+
+export interface ChainSettings {
+  verifiers: SettingsForVerifier
+}
+
+export type SettingsForChains = Record<string, ChainSettings>
+
+export interface ContractVerificationSettings {
+  chains: SettingsForChains
+}
