+++ apps/contract-verification/src/app/types/index.ts
@@ -1,1 +1,3 @@
 export * from './ThemeType'
+export * from './SettingsTypes'
+export * from './VerificationTypes'
